pyproject.toml
[tool.poetry]
name = "atleta"
version = "0.1.0"
description = ""
authors = ["Fuffes <Fuffes05@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.5.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
